Add files via upload
ROZTEST shows current analog measurement AND min and max values, which latch until a more significant value is read. Millis are also shown with each displayed line of measurement, all for easier line-by-line comparison of data.
Original code was for analog measurements from a single-axis accelerometer on an Arduino Uno and read via the serial terminal.